feat: implement the `analyzer` command.
This commit introduces the `sprocket analyzer` command that can be used to run
sprocket as an LSP server.

The command currently takes two options:

* `--lint`: whether or not to enable lints in analysis.
* `--stdio` (currently required): use stdio as the transport channel for the
  server.

This commit also updates all dependencies to latest and adds `Cargo.lock` to
the repository for consistent builds, which is desirable for bin projects like
`sprocket`.