feat: add the `sprocket format` command.
This commit adds the `format` command to `sprocket`.

The command will format a single WDL document.

Also updated the existing commands to the latest WDL API and corrected the
environment variable name for publishing crates.

Included the analysis rules in `sprocket explain`; the list of rules is now
sorted by id.