feat: add support for formatting directories.
This commit adds support for formatting directories.

By default, the `format` command will attempt to print a single formatted
document to stdout.

If the `--overwrite` option is specified, a directory may be specified and all
of the WDL documents recursively found in the directory will be overwritten
with the formatted contents.