feat: introduces incremental command line parsing
This commit is the second half of the changes needed to land incremental command
line parsing into `sprocket`. The first half is at stjude-rust-labs/wdl#430.

As indicated in the changes there, this commit also has the side effect of
removing the `wdl` command line tool and putting all of the subcommands in
`sprocket` instead. As such, you'll see many commands be modified to use the
code created in the new `wdl-cli` package _OR_ brand new subcommands (like
`run`).