feat: integrate engine configuration into `sprocket.toml`.
This commit adds a `run` configuration (section) to `sprocket.toml`.

The contents of the configuration is a flattened `wdl-engine` configuration.

This allows users to have a single configuration for controlling sprocket.

As a consequence, the `--config` option that was specific to `sprocket run` was
removed. As there aren't many users of `sprocket run` currently, I opted for
deleting the option rather than deprecating it.

Users should migrate the contents of their engine configurations to a
`sprocket.toml`.

Note that the configuration format for the engine has also changed in
`wdl-engine`: multiple backends may now be defined, which also allows sprocket
to merge them from multiple configuration files.

Additionally, `sprocket run` has been changed so that passing it a file with a
single task and no workspace runs without having to specify the `--name`
option.