fix: do not flatten engine configuration as optional.
Currently we flatten the engine configuration into the `run` configuration as
`Option`. This has the unintended consequence of turning errors in
deserializing a field within the engine configuration into a `None` for the
entire engine configuration.

The result is implicitly using a default engine configuration when an engine
configuration field fails to deserialize due to a call to `unwrap_or_default`.

The fix is to not make `engine` in the run configuration `Option`.