chore: update `indicatif`, Cargo.lock, and enforce lockfile in CI
I initially noticed the `Cargo.lock` file is out of date on `main`, and in trying to address that
realized that we inadvertently grew a diamond dependency problem around `indicatif` following a
release of `tracing-indicatif`. This PR fixes that immediate issue, and also adds CI enforcement for
the lockfile so that such problems will be caught earlier.