feat: add configurable unrecognized version fallback behavior (#475)
Implements the `wdl` side of #464; a new `sprocket` release will still be required to fully close out that issue.

The primary change here is a move of the recognized version enforcement from the initial parse in `wdl-grammar` to the code in `wdl-analysis` that consumes the parse. The behavior upon finding an unrecognized version is now also configurable, with a configurable-severity diagnostic being emitted in case the fallback is used. Other notable changes include:

- `Analyzer` now takes a general-purpose `Config` struct which contains the previous `DiagnosticsConfig` struct as a field.
- `wdl-ast` interfaces now include a version fallback option in order to return an `Ast` variant consistent with the analyzer's configured fallback option.
- `wdl-lsp`'s configuration is now hardcoded to fall back to the default WDL version (1.2 as of this change) with warning-level severity for the associated diagnostic. Additional configurability is planned in #517.
- The `wdl-analysis` test suite now looks for `config.toml` in test definition directories and runs the test under that configuration if found. This is a fairly crude mechanism at this point—suitable for testing, but `sprocket`'s config files are the better option for user-facing config.
- New tests have been added that exercise the fallback behavior with and without configured diagnostics.