chore: update CLI tests to make output files baseline optional.
This commit changes the CLI tests so that the `outputs` directory is now
optional.

If the directory exists, the test will compare it against the resulting working
directory. This is primarily needed for commands line `sprocket format` that
modify files in the working directory.

For the `sprocket run` commands, the intermediate files aren't as interesting
as the stdout/stderr output of sprocket.