fix: use a `tmp` location to allow for hardlinking.
Currently the file transferer uses the system tmp directory for linking files
in the download cache to a location that can be used by Sprocket.

On many systems, though, the location is likely to be a different file system
than where `sprocket` is configured to store its download cache; this prevents
hardlinks and forces a fallback to a copy.

This fixes the transferer such that it uses a tmp directory within the cache
root to maximum the success of creating a hardlink to the cache file.