fix: improve error message from downloaded WDL source files.
This commit improves the error message encountered when the server returns an
error for a remote WDL source file to include the URI of the file being
downloaded.

While failure to resolve an import already contains the path to the URI in
error messages, if the top-level source file is a remote URL and the server
returns an error, an unhelpful message of `error: server returned HTTP status
404 Not Found` without context.

This can be easily observed today via `sprocket run https://google.com/404`.