fix: fix a `debug!` log which causes a panic.
The problem occurs when:

1. Logging with DEBUG level (i.e. `-vv`).
2. Using the TES backend.
3. Using inputs that are remote URLs.

There's a `debug!` log message in task evaluation that logs what inputs have
been downloaded and mapped to.

The logging assumes that if the input is remote, it has been downloaded (i.e.
localized).

But that's not the case with the TES backend; the input can be remote and _not_
downloaded because task execution occurs remotely.

The fix is to properly add a match case to handle it.