feat: support `glob` on remote working directories.
This commit adds support to the `glob` function so that it can work with remote
working directories (i.e. with the TES backend).

It updates to the latest `cloud-copy` to use the `walk` function for
enumerating the "child" URLs of the working directory and uses `globset` to
filter the list based on the glob.

Additionally, this also fixes determining if a remote URL "exists", which is
required for properly handling optional `File` and `Directory` values.

The `crankshaft` crate has been updated to the latest version to improve the
TES backend polling performance.