feature: add per-queue CPU and memory limits for the LSF backend
For now, these are manually configured based on the user's understanding of each queue's
behavior. LSF tooling doesn't offer a straightforward way to ask "what's the most CPU I can schedule
on this queue?", but we could probably stitch a reasonable heuristic together in the future by
joining `bqueues` and `bhosts` output.

Adding these limits also means that the entire `wdl-engine/tasks` suite can run and pass in a
suitable HPC environment. To that end, this change also extends that suite to check whether LSF and
Apptainer tooling is available, and to run the tests with that backend if so. For the moment, this
facilitates manual testing of the HPC backend rather than being hooked into CI, but it is an
improvement over the previous state-of-the-art which was me continually rebasing a set of
HPC-specific changes to push over to the HPC on a separate in-situ branch.