revise: improve configurability of integration test suites and make WDL 1.3 flag specific to that version (#435)
These changes are primarily aimed at improving the suitability of the test suite for execution in HPC environments in service of #434, but also give `wdl-engine` test authors more flexibility to vary test configurations.

## Engine config from environment

The `SPROCKET_TEST_ENGINE_CONFIG` environment variable can be set to a file path that will be read in place of the default engine configuration(s) whenever a task or workflow will be executed by a test. Many of our tests will, by default, run using the local backend and the Docker backend. By setting the environment variable to an alternate configuration like `test-configs/lsf-apptainer-engine.toml`, these default backends will be omitted and only the specified backend will be used.

So far, this capability has been added to the `sprocket/cli/run` suite and the `wdl-engine/{tasks,workflows}` suites, but it is portable to any test suite which may require the use of `wdl-engine` evaluation in the future.

This replaces the recently-added mechanism that checks for LSF and Apptainer binaries on the PATH and performs a similar config substitution. This approach quickly fell over when encountering HPC environments with multiple scheduler binaries on the PATH.

## Config overrides instead of replacements

To make this change tractable in combination with the existing ability to override engine configs on a test-by-test basis via `config.toml` or `config.linux.toml` in the test root, the latter capability has been changed to be a means to modify the base config rather than replacing it in full.

Instead of a full config in TOML format, the test runners for the `tasks` and `workflows` suites look for a `config-override.yaml` file and apply it using [Figment's `merge` semantics](https://docs.rs/figment/latest/figment/struct.Figment.html#conflict-resolution). YAML was chosen due to its ability to override `Some(_)` options with `None`; TOML does not have a `null` so this is not easily expressible.

## Config overrides for `Analysis`

#424 added configs for the `Analysis` step of the `wdl-engine` tests beyond the defaults. Taking advantage of the new capability to more-surgically override config fields for individual tests, this change makes the analysis config overridable using the same `config-override.yaml` as the engine config.

## Code consolidation in the `wdl-engine` integration tests

Much of the logic in the `tasks` and `workflows` suites was redundant, and has been combined in a `common` module.

## Increased specificity of the feature flag enabling WDL 1.3 features

While integrating the new override for the analysis config, I noticed that this new feature flag enables all potential future versions, not just WDL 1.3. This scope appeared to be unintentionally broad, so I narrowed it to 1.3-only with the intent that it can be deprecated and removed once 1.3 stabilizes.

## Shell expansion on `apptainer_images_dir` config flag

This is needed in order to specify a safe place for these cache directories on HPCs where `$TMPDIR` cannot be reliably used across hosts. This will likely be obviated by #360 but will be useful in the meantime.

## Fixed a quoting bug when passing environment variables to Apptainer

How about that, improving the test suite found a bug!