feat: implement "slow" vs "fast" failure modes.
This commit implements both a "slow" and a "fast" failure mode in Sprocket.

The "slow" failure mode waits for currently executing tasks to complete when an
error is encountered or the user interrupts evaluation.

The "fast" failure mode immediately cancels the currently executing tasks and
waits for cancellation to complete.

The previous implementation of Sprocket behaved like the "fast" failure mode;
the new default is "slow" and can be changed via a setting in `sprocket.toml`.

Additionally, Ctrl-C handling has been reimplemented to support multiple Ctrl-C
invocations depending on the failure mode.