fix: improve portability of Apptainer scripts (#442)
Thanks to the efforts of brave early adopters, we were reminded that `/usr/bin/env` is a more common location for `env` than `/bin/env`, and discovered that the `--cwd` argument to `apptainer exec` is an alias for `--pwd` that was introduced in Apptainer 1.3. This change switches both of these to the variation that appears more likely to be consistent across platforms. If we encounter problems with these, it may make sense to make these configurable in the future.