wip: rewrite `wdl_engine::path` URL helpers for better safety
The names of these functions were a little misleading. Notably, `is_url()` returned `false` for URLs
that had unsupported schemes, and examining some call sites makes me think that `!is_url()` was
often interpreted to mean "this *is* a file path".

I renamed the functions, and rewrote the implementations to perform the `Url` parsing first. In a
subsequent pass, I will look at more of the call sites and see what I can do to remove redundant
parsing: if we have a URL, we should parse it as soon as possible and keep it in that form rather
than keeping it as a string that gets repeatedly re-parsed.