refactor: make CLI input processing less dependent on mutation (#460)
This is an intermediate step in a larger refactoring effort to make the interface to `wdl-engine` smaller and more consistent. The next step of the refactoring is likely to be quite large, so even though this is not a complete set of desired changes, we have decided to land it in advance of that to avoid the size of any one PR getting out of control.

Each commit on this branch ought to work independently, and the `wip` commit entries should give a clue as to what's going on. I'll call out a few larger changes at the top level though:

- `ensure_paths_exist()` has been renamed to `resolve_paths()`, and now works in a functional style of returning a new `Value` rather than mutating one in place. On its face, this is unsatisfying as a number of values will get copied unnecessarily, but since this is logic that runs once on a finite set of initial inputs on startup, the cost should be negligible compared to the benefits in code clarity.
- One layer of `join_paths()` has been removed via inlining, though `Task::join_paths()` and `Workflow::join_paths()` remain as functions that mutate inputs. This will be addressed in subsequent work.
- Helper functions in `wdl_engine::path` have been rewritten to parse strings into URLs early and often, rather than relying on our own partial implementations of URL parsing. I am planning in a subsequent step to reduce the number of redundant parses this introduces by using URLs in the representation of values earlier in the evaluation process.
- Instead of "inputs" or anonymous tuples, I've begun using "invocation" to describe the combinations of entrypoint, input values, and origin paths, though this transformation is incomplete as of this PR.

Co-authored-by: Clay McLeod <3411613+claymcleod@users.noreply.github.com>
Co-authored-by: Peter Huene <peter@huene.dev>