fix: prevent the loss of error context from Sprocket commands.
This commit fixes Sprocket commands such that they don't erase error
context by creating a new error derived from an existing error.

The commands now return a `CommandResult` which has an error that can be
a single `anyhow::Error` or multiple shared errors from analysis.
Previously it was returning `anyhow::Result`, which meant the shared
errors had to convert to an owned `anyhow::Error`, which causes the loss
of context.

Fixes #181.