feat: new `wdl-engine` evaluation entrypoints and Apptainer cache management (#463)
Closes #360, and provides a small measure of unification for the `wdl-engine` evaluator entrypoints.

For implementors, there is now a `TopLevelEvaluator` struct that's available throughout task and workflow evaluation and is suitable for tracking once-per-run data like the Apptainer image cache. Previously, only the root of the current task or workflow's execution was available, which would have led to no image reuse at all if it had been used for Apptainer.