fix: improve digest caching.
This commit improves digest caching whereby the calculation of a local
directory digest will cache any decedent file or directory digest that is
calculated.

The fix should help improve performance whereby a task creates a large file in
its working directory and a direct reference to the file is later used by
another task.

Without this fix, the above would cause the referenced file to be digested
twice: once as a decedent of the directory and again for the direct reference
to the file.

Note: due to the change in how local directory digests are calculated, this fix
will cause call cache misses for any entry that references a local directory.