fix: buffer more events for Sprocket's progress statistics.
The events channel capacity was defaulting to `100`, which was overly
conservative and caused users to observe Sprocket no long displaying
progress statistics when a lot of tasks have been concurrently queued.

This commit changes the default to `5000`, which will allocate more
memory but make reaching the capacity much more unlikely.

Additionally, this adds the `run.events_capacity` setting in
`sprocket.toml` for users that want even more events to be buffered.