refactor: change `evaluate_task` to return "failed" evaluated tasks.
This commit refactors `TopLevelEvaluator::evaluate_task` to return the
evaluated task when it has an unacceptable error code.

Previously we were treating it as an evaluation error, which internally
gives no context other than the error message as to what the task
execution result was; this is needed for implementing assertions in
`sprocket test`).

Calling `into_outputs` on the returned evaluated task will return an
error if it had an unacceptable exit code; the error will have an
associated task execution result that stores the task's exit code.

This also includes the following fixes:

* Only attempt to download `stderr` of the failed task when the maximum
  retry is exceeded.
* We now properly invoke the execution backend cleanup when the task had
  an unacceptable exit code.