fix: add dependency edges to scatter variables correctly.
This commit fixes a bug in forming the dependency graph for workflow
evaluation.

If an expression from within a scatter statement referenced the scatter
variable, but private declaration from outside of the scatter exists
with the same name, an incorrect edge was added to the declaration
instead of the scatter variable.

The fix is to check for the presence of the scatter variable before
checking for other names; this allows a scatter variable to effectively
shadow a private declaration of the same name made from outside of the
scatter statement. It remains illegal to shadow a scatter variable with
a declaration of the same name from _within_ the scatter's scope.

Fixes #508.